Move battle and map room classes
