+++ roomeditor/forms.py
@@ -7,8 +7,8 @@ class RoomForm(forms.Form):
     ROOM_CLASS_CHOICES = [
         ("typeclasses.rooms.Room", "Room"),
         ("typeclasses.rooms.FusionRoom", "Fusion Room"),
-        ("typeclasses.battleroom.BattleRoom", "Battle Room"),
-        ("typeclasses.maproom.MapRoom", "Map Room"),
+        ("typeclasses.rooms.BattleRoom", "Battle Room"),
+        ("typeclasses.rooms.MapRoom", "Map Room"),
     ]
 
     ROOM_CLASS_HELP = (
